cleaned up some code smells
