+++ app/server.js
@@ -1,10 +1,8 @@
 // Backend @authors: Jess Cannarozzo & Karla Martins-Spuldaro
 const express = require('express');
 const app = express();
-const request = require('request');
-const quotesURL = "https://api.forismatic.com/api/1.0/?method=getQuote&lang=en&format=jsonp&jsonp=?";
+const quoteModule = require('./quotes-module.js');
 
-const quoteLen = 60;
 
 //routes
 app.get('/', (req, res) => {
@@ -13,7 +11,7 @@ app.get('/', (req, res) => {
 
 // output: passwordObj with generated quote password, quote, author, quote link
 app.get('/quotes', (req, res) => { 
-    getQuote(res);
+    quoteModule.getQuote(res);
 });
 
 app.get('/colours', (req,res) => { 
@@ -24,55 +22,5 @@ app.get('/credits', (req,res) => {
     res.sendStatus(200);
 });
 
-//methods
-//input: quote (string)
-//output: create password from first letter of each word in the quote
-function createQuotePW(quote) {
-    if (quote) {
-        const quotePW = quote.match(/\b(\w)/g).join('');
-        
-        return quotePW;
-    }
-}
-
-// gets quote from quotesURL, returns quote to res
-function getQuote(cb) {
-    request.get({
-        url: quotesURL
-    }, (error, message,body) => {
-        body = formatQuote(body);
-        if (!error && message.statusCode === 200 && JSON.parse(body).quoteText.length < quoteLen) {          
-        //    body = formatQuote(body);
-            let quote = {};
-            
-            try {
-                quote = JSON.parse(body);
-            } catch(err) {
-                console.log(err);
-            }
-
-            console.log(quote.quoteText);
-                        
-            
-            const passwordObj = {
-                pw: createQuotePW(quote.quoteText),
-                quote: quote.quoteText,
-                author: quote.quoteAuthor,
-                link: quote.quoteLink
-            };
-            return cb.send(passwordObj)
-        } else {
-            getQuote(cb);
-        }
-    });
-}
-
-// removes odd characters from quote
-function formatQuote(body) {
-    if (body.charAt(0) === '?') { //format
-        // console.log(body);
-        return body.substring(2,body.length-1);
-    }
-}
 
 app.listen(3000, () => console.log("We're live on 3000! In style."));
